Add initial implementation to add new actors
This requires setting up the NewActor component
to use Auth0 to both (a) get the access token
for the current user and (b) pass that token
into the endpoint call.

This also means moving the NewActor component
to be a functional component so I could take
advantage of react hooks.

I left the defaults in for testing.  I will
remove once I can delete actors more easily
or have a way to pre-populate from a list
for show.