Backend just serves the API surface now
The backend is live at https://nickanthony-casting-agency.herokuapp.com/

The frontend is live at https://nickanthony-casting-agency-c.herokuapp.com/

This change reverts out the frontend being served by the backend
because react router breaks in that case.  In effect, only the backend
serves the index route, then react router handles the further redirects.
However, this breaks if you reload the page because the backend
gets queried again with a URL generated by react router.  This
causes a 404.  So I separated the instances so we don't have this
problem.